# internal/apidocs/gen/swagger.json
{
    "components": {
        "schemas": {
...
            "server.hostStatusMemory": {
                "properties": {
                    "free": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "used": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "server.hostStatusResponse": {
                "properties": {
                    "hostname": {
                        "type": "string"
                    },
                    "load_average": {
                        "items": {
                            "type": "number"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "memory": {
                        "$ref": "#/components/schemas/server.hostStatusMemory"
                    },
                    "uptime": {
                        "$ref": "#/components/schemas/server.hostStatusUptime"
                    }
                },
                "type": "object"
            },
            "server.hostStatusUptime": {
                "properties": {
                    "boot_time": {
                        "type": "string"
                    },
                    "formatted": {
                        "type": "string"
                    },
                    "seconds": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "server.hostUptimeLoadAverages": {
                "properties": {
                    "15min": {
                        "type": "number"
                    },
                    "1min": {
                        "type": "number"
                    },
                    "5min": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
            "server.hostUptimeResponse": {
                "properties": {
                    "boot_time": {
                        "type": "string"
                    },
                    "load_averages": {
                        "$ref": "#/components/schemas/server.hostUptimeLoadAverages"
                    },
                    "uptime_formatted": {
                        "type": "string"
                    },
                    "uptime_seconds": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "server.powerRequest": {
                "properties": {
                    "confirm": {
                        "type": "boolean"
                    },
                    "emergency": {
                        "description": "Halt only: emergency acknowledgement (halt skips graceful shutdown entirely)",
                        "type": "boolean"
                    },
                    "grace_period": {
                        "description": "Seconds before the action fires (Unix rounds up to minutes)",
                        "type": "integer"
                    },
                    "message": {
                        "description": "Warning broadcast to logged-in users where the platform supports one",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "server.powerTaskResponse": {
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "grace_period": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "task_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "The swag-generated half of the Agent API v1 document. Serve-time authority lives in internal/apidocs: the hand-maintained fragment (openapi.json) owns info/tags/components/security and every not-yet-migrated path, and this generated document's paths merge into it.",
        "title": "Agent API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/system/host/halt": {
            "post": {
                "description": "Minimum role: admin. No grace period, no graceful shutdown — the emergency stop. Requires BOTH confirm:true and emergency:true. 503 when host_power.enabled is false.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/server.powerRequest",
                                        "summary": "request",
                                        "description": "Power action body (confirm and emergency required)"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Power action body (confirm and emergency required)",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Halt task created"
                    },
                    "400": {
                        "description": "Missing confirmation or emergency acknowledgement"
                    },
                    "503": {
                        "description": "Host power management is disabled in configuration"
                    }
                },
                "summary": "Immediately halt the host",
                "tags": [
                    "System Host Management"
                ]
            }
        },
        "/system/host/poweroff": {
            "post": {
                "description": "Minimum role: admin. Manual intervention required to restart the machine. Same body rules as shutdown; Windows makes no shutdown/poweroff distinction. 503 when host_power.enabled is false.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/server.powerRequest",
                                        "summary": "request",
                                        "description": "Power action body (confirm required)"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Power action body (confirm required)",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Poweroff task created"
                    },
                    "400": {
                        "description": "Missing confirmation or invalid parameters"
                    },
                    "503": {
                        "description": "Host power management is disabled in configuration"
                    }
                },
                "summary": "Power off the host",
                "tags": [
                    "System Host Management"
                ]
            }
        },
        "/system/host/restart": {
            "post": {
                "description": "Minimum role: admin. Same body rules as shutdown. 503 when host_power.enabled is false.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/server.powerRequest",
                                        "summary": "request",
                                        "description": "Power action body (confirm required)"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Power action body (confirm required)",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Restart task created"
                    },
                    "400": {
                        "description": "Missing confirmation or invalid parameters"
                    },
                    "503": {
                        "description": "Host power management is disabled in configuration"
                    }
                },
                "summary": "Restart the host",
                "tags": [
                    "System Host Management"
                ]
            }
        },
        "/system/host/shutdown": {
            "post": {
                "description": "Minimum role: admin. Queues a task running the platform shutdown command (Windows shutdown /s, Unix shutdown -h). confirm:true is required. The agent needs the OS privilege the command itself demands — a refusal fails the task honestly. 503 when host_power.enabled is false.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/server.powerRequest",
                                        "summary": "request",
                                        "description": "Power action body (confirm required)"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Power action body (confirm required)",
                    "required": true
                },
                "responses": {
                    "202": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.powerTaskResponse"
                                }
                            }
                        },
                        "description": "Shutdown task created"
                    },
                    "400": {
                        "description": "Missing confirmation or invalid parameters"
                    },
                    "503": {
                        "description": "Host power management is disabled in configuration"
                    }
                },
                "summary": "Shut down the host",
                "tags": [
                    "System Host Management"
                ]
            }
        },
        "/system/host/status": {
            "get": {
                "description": "Minimum role: viewer. Uptime, load averages, and memory. No runlevel or reboot-required tracking — init concepts absent on this platform trio. 503 when host_power.enabled is false.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.hostStatusResponse"
                                }
                            }
                        },
                        "description": "System status"
                    },
                    "503": {
                        "description": "Host power management is disabled in configuration"
                    }
                },
                "summary": "Host system status",
                "tags": [
                    "System Host Management"
                ]
            }
        },
        "/system/host/uptime": {
            "get": {
                "description": "Minimum role: viewer. 503 when host_power.enabled is false.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.hostUptimeResponse"
                                }
                            }
                        },
                        "description": "Uptime information"
                    },
                    "503": {
                        "description": "Host power management is disabled in configuration"
                    }
                },
                "summary": "Host uptime",
                "tags": [
                    "System Host Management"
                ]
            }
        },
...
    },
    "openapi": "3.1.0"
}